% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_app.R
\name{chat_app}
\alias{chat_app}
\alias{chat_mod_ui}
\alias{chat_mod_server}
\title{Open a live chat application in the browser}
\usage{
chat_app(client, ...)

chat_mod_ui(id, client, ...)

chat_mod_server(id, client)
}
\arguments{
\item{client}{A chat object created by \pkg{ellmer}, e.g.
\code{\link[ellmer:chat_openai]{ellmer::chat_openai()}} and friends.}

\item{...}{In \code{chat_app()}, additional arguments are passed to
\code{\link[shiny:shinyApp]{shiny::shinyApp()}}. In \code{chat_mod_ui()}, additional arguments are passed to
\code{\link[=chat_ui]{chat_ui()}}.}

\item{id}{The chat module ID.}
}
\value{
A \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
Create a simple Shiny app for live chatting using an \link[ellmer:Chat]{ellmer::Chat} object.
Note that these functions will mutate the input \code{client} object as
you chat because your turns will be appended to the history.
}
\section{Functions}{
\itemize{
\item \code{chat_app()}: A simple Shiny app for live chatting.

\item \code{chat_mod_ui()}: A simple chat app module UI.

\item \code{chat_mod_server()}: A simple chat app module server.

}}
\examples{
\dontrun{
client <- ellmer::chat_claude()
chat_app(client)
}

}
