% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_enable_bookmark.R
\name{chat_enable_bookmark}
\alias{chat_enable_bookmark}
\title{Add Shiny bookmarking for shinychat}
\usage{
chat_enable_bookmark(
  id,
  client,
  ...,
  update_on_input = TRUE,
  update_on_response = TRUE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{The ID of the chat element}

\item{client}{The \pkg{ellmer} LLM chat client.}

\item{...}{Used for future parameter expansion.}

\item{update_on_input}{A logical value determines if the bookmark should be updated when the user submits a message. Default is \code{TRUE}.}

\item{update_on_response}{A logical value determines if the bookmark should be updated when the response stream completes. Default is \code{TRUE}.}

\item{session}{The Shiny session object}
}
\value{
Returns nothing (\code{invisible(NULL)}).
}
\description{
Adds hooks to the Shiny chat given the LLM client.

If either \code{update_on_input} or \code{update_on_response} is \code{TRUE}, the Shiny
App's bookmark will be automatically updated without showing a modal to the
user.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(shiny)
library(bslib)
library(shinychat)

ui <- page_fillable(
  chat_ui("chat", fill = TRUE)
)

server <- function(input, output, session) {
  chat_client <- ellmer::chat_ollama(
    system_prompt = "Important: Always respond in a limerick",
    model = "qwen2.5-coder:1.5b",
    echo = TRUE
  )
  # Update bookmark to chat on user submission and completed response
  chat_enable_bookmark("chat", chat_client)

  observeEvent(input$chat_user_input, {
    stream <- chat_client$stream_async(input$chat_user_input)
    chat_append("chat", stream)
  })
}

# Enable bookmarking!
shinyApp(ui, server, enableBookmarking = "server")
\dontshow{\}) # examplesIf}
}
