% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat_ui}
\alias{chat_ui}
\title{Create a chat UI element}
\usage{
chat_ui(
  id,
  placeholder = "Enter a message...",
  width = "min(680px, 100\%)",
  height = "auto",
  fill = TRUE,
  ...
)
}
\arguments{
\item{id}{The ID of the chat element}

\item{placeholder}{The placeholder text for the chat's user input field}

\item{width}{The CSS width of the chat element}

\item{height}{The CSS height of the chat element}

\item{fill}{Whether the chat element should try to vertically fill its
container, if the container is
\href{https://rstudio.github.io/bslib/articles/filling/index.html}{fillable}}

\item{...}{Extra HTML attributes to include on the chat element}
}
\value{
A Shiny tag object, suitable for inclusion in a Shiny UI
}
\description{
Create a chat UI element
}
