% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-stream.R
\name{output_markdown_stream}
\alias{output_markdown_stream}
\title{Create a UI element for a markdown stream.}
\usage{
output_markdown_stream(id, ..., content = "", content_type = "markdown")
}
\arguments{
\item{id}{A unique identifier for this markdown stream.}

\item{...}{Extra HTML attributes to include on the chat element}

\item{content}{Some content to display before any streaming occurs.}

\item{content_type}{The content type. Default is \code{"markdown"} (specifically,
CommonMark). Other supported options are:
* \code{"html"}: for rendering HTML content.
* \code{"text"}: for plain text.
* \code{"semi-markdown"}: for rendering markdown, but with HTML tags escaped.}
}
\value{
A shiny tag object.
}
\description{
Creates a UI element for a \code{\link[=markdown_stream]{markdown_stream()}}. A markdown stream can be
useful for displaying generative AI responses (outside of a chat interface),
streaming logs, or other use cases where chunks of content are generated
over time.
}
\seealso{
\code{\link[=markdown_stream]{markdown_stream()}}
}
